pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "venture-image"
version = "0.1.0"
description = "Modular image/file management backend (FastAPI) with pluggable modules via entry points."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Saj Rags" }]

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.5",
  "pillow>=10.0",
  "imagehash>=4.3",
  "tqdm>=4.66",
  "exifread>=3.0.0",
  "geopy>=2.4",
  "requests>=2.32",
  "typer>=0.12",
]

# Nice-to-haves during dev with uv's dependency groups
[dependency-groups]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.5", "mypy>=1.10", "httpx>=0.27"]

[project.optional-dependencies]
# Kept for compatibility with non-uv tooling; uv will prioritize dependency-groups.
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.5", "mypy>=1.10", "httpx>=0.27"]

[project.urls]
Source = "https://example.com/venture-image"

# --- Entry-point based module discovery ---
[project.entry-points."vi_app.modules"]
cleanup = "vi_app.modules.cleanup.router:router"
convert_images = "vi_app.modules.convert_images.router:router"
# Add others as you build them:
# dedup = "vi_app.modules.dedup.router:router"
# sort = "vi_app.modules.sort.router:router"

[tool.hatch.build.targets.wheel]
packages = ["src/vi_app"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
